# docs/openapi.yaml
openapi: 3.1.0
info:
  title: Fuse AI System Manager API
  version: 0.1.0
  description: >
    Fuse is an open-source AI system manager exposing three compatibility layers
    (Ollama, OpenAI, Anthropic) plus native system endpoints.  A single Rust
    binary replaces Ollama + vLLM and runs on any device from Raspberry Pi to
    data center.
  license:
    name: Apache-2.0
    url: https://www.apache.org/licenses/LICENSE-2.0
  contact:
    name: Fuse Maintainers
servers:
  - url: http://localhost:11434
    description: Local development server (default Ollama-compatible port)
...
components:
  securitySchemes:
    ApiKeyHeader:
      type: apiKey
      in: header
      name: X-API-Key
      description: API key passed via the X-API-Key header.
    BearerAuth:
      type: http
      scheme: bearer
      description: Bearer token passed via the Authorization header.
  headers:
    X-RateLimit-Remaining:
      description: Number of requests remaining in the current window.
      schema:
        type: integer
    X-RateLimit-Reset:
      description: UTC epoch seconds when the rate-limit window resets.
      schema:
        type: integer
  schemas:
...
    ErrorResponse:
      type: object
      required: [error_code, message]
      properties:
        error_code:
          type: string
          description: Machine-readable error code.
          examples: ["rate_limit_exceeded"]
        message:
          type: string
          description: Human-readable error description.
          examples: ["Rate limit exceeded. Please try again later."]
        remediation:
          type: string
          description: Suggested action to resolve the error.
          examples: ["Wait 30 seconds and retry."]

    # ------------------------------------------------------------------
    # Ollama schemas
    # ------------------------------------------------------------------
...
    OllamaEmbeddingsRequest:
      type: object
      required: [model, prompt]
      properties:
        model:
          type: string
        prompt:
          type: string
        options:
          type: object
    OllamaEmbeddingsResponse:
      type: object
      required: [embedding]
      properties:
        embedding:
          type: array
          items:
            type: number
            format: float

    # ------------------------------------------------------------------
    # OpenAI schemas
    # ------------------------------------------------------------------
    OpenAIChatMessage:
      type: object
      required: [role]
      properties:
        role:
          type: string
          enum: [system, user, assistant, tool]
        content:
          type: string
        tool_calls:
          type: array
          items:
            $ref: "#/components/schemas/ToolCall"
        tool_call_id:
          type: string
        name:
          type: string
    ToolCall:
      type: object
      required: [id, type, function]
      properties:
        id:
          type: string
        type:
          type: string
          enum: [function]
        function:
          $ref: "#/components/schemas/ToolCallFunction"
    ToolCallFunction:
      type: object
      required: [name, arguments]
      properties:
        name:
          type: string
        arguments:
          type: string
          description: JSON-encoded function arguments.
    ToolDefinition:
      type: object
      required: [type, function]
      properties:
        type:
          type: string
          enum: [function]
        function:
          $ref: "#/components/schemas/FunctionDefinition"
    FunctionDefinition:
      type: object
      required: [name]
      properties:
        name:
          type: string
        description:
          type: string
        parameters:
          type: object
          description: JSON Schema for function parameters.
    ResponseFormat:
      type: object
      required: [type]
      properties:
        type:
          type: string
          enum: [text, json_object]
    ChatCompletionRequest:
      type: object
      required: [model, messages]
      properties:
        model:
          type: string
          examples: ["gpt-4"]
        messages:
          type: array
          items:
            $ref: "#/components/schemas/OpenAIChatMessage"
        temperature:
          type: number
          format: double
          minimum: 0
          maximum: 2
        top_p:
          type: number
          format: double
          minimum: 0
          maximum: 1
        max_tokens:
          type: integer
          format: uint32
        stream:
          type: boolean
          default: false
        tools:
          type: array
          items:
            $ref: "#/components/schemas/ToolDefinition"
        tool_choice:
          description: '"auto", "none", "required", or a specific function object.'
        response_format:
          $ref: "#/components/schemas/ResponseFormat"
        stop:
          description: String or array of strings.
        n:
          type: integer
          format: uint32
          default: 1
        presence_penalty:
          type: number
          format: double
        frequency_penalty:
          type: number
          format: double
        seed:
          type: integer
          format: uint64
    Usage:
      type: object
      required: [prompt_tokens, completion_tokens, total_tokens]
      properties:
        prompt_tokens:
          type: integer
          format: uint32
        completion_tokens:
          type: integer
          format: uint32
        total_tokens:
          type: integer
          format: uint32
    ChatCompletionChoice:
      type: object
      required: [index, message, finish_reason]
      properties:
        index:
          type: integer
          format: uint32
        message:
          $ref: "#/components/schemas/OpenAIChatMessage"
        finish_reason:
          type: string
          enum: [stop, length, tool_calls, content_filter]
    ChatCompletionResponse:
      type: object
      required: [id, object, created, model, choices, usage]
      properties:
        id:
          type: string
          examples: ["chatcmpl-abc123"]
        object:
          type: string
          enum: [chat.completion]
        created:
          type: integer
          format: int64
        model:
          type: string
        choices:
          type: array
          items:
            $ref: "#/components/schemas/ChatCompletionChoice"
        usage:
          $ref: "#/components/schemas/Usage"
    ChatCompletionDelta:
      type: object
      properties:
        role:
          type: string
        content:
          type: string
        tool_calls:
          type: array
          items:
            $ref: "#/components/schemas/ToolCall"
    ChatCompletionChunkChoice:
      type: object
      required: [index, delta]
      properties:
        index:
          type: integer
          format: uint32
        delta:
          $ref: "#/components/schemas/ChatCompletionDelta"
        finish_reason:
          type: string
          enum: [stop, length, tool_calls, content_filter, null]
    ChatCompletionChunk:
      type: object
      required: [id, object, created, model, choices]
      properties:
        id:
          type: string
        object:
          type: string
          enum: [chat.completion.chunk]
        created:
          type: integer
          format: int64
        model:
          type: string
        choices:
          type: array
          items:
            $ref: "#/components/schemas/ChatCompletionChunkChoice"
    EmbeddingRequest:
      type: object
      required: [model, input]
      properties:
        model:
          type: string
          examples: ["text-embedding-ada-002"]
        input:
          description: A string or array of strings to embed.
        encoding_format:
          type: string
          enum: [float, base64]
    EmbeddingData:
      type: object
      required: [object, embedding, index]
      properties:
        object:
          type: string
          enum: [embedding]
        embedding:
          type: array
          items:
            type: number
            format: float
        index:
          type: integer
          format: uint32
    EmbeddingUsage:
      type: object
      required: [prompt_tokens, total_tokens]
      properties:
        prompt_tokens:
          type: integer
          format: uint32
        total_tokens:
          type: integer
          format: uint32
    EmbeddingResponse:
      type: object
      required: [object, data, model, usage]
      properties:
        object:
          type: string
          enum: [list]
        data:
          type: array
          items:
            $ref: "#/components/schemas/EmbeddingData"
        model:
          type: string
        usage:
          $ref: "#/components/schemas/EmbeddingUsage"
...
security:
  - ApiKeyHeader: []
  - BearerAuth: []
paths:
...
  /api/embeddings:
    post:
      operationId: ollamaEmbeddings
      tags: [Ollama]
      summary: Generate embeddings
      description: Generate an embedding vector for the given prompt.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/OllamaEmbeddingsRequest"
      responses:
        "200":
          description: Embedding vector.
          headers:
            X-RateLimit-Remaining:
              $ref: "#/components/headers/X-RateLimit-Remaining"
            X-RateLimit-Reset:
              $ref: "#/components/headers/X-RateLimit-Reset"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OllamaEmbeddingsResponse"

  # ====================================================================
  # OpenAI endpoints
  # ====================================================================
  /v1/chat/completions:
    post:
      operationId: openaiChatCompletions
      tags: [OpenAI]
      summary: Chat completions
      description: >
        Create a chat completion.  When `stream` is true the response is
        Server-Sent Events (text/event-stream) with `ChatCompletionChunk`
        objects; the stream ends with a `data: [DONE]` sentinel.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ChatCompletionRequest"
      responses:
        "200":
          description: Chat completion or SSE stream.
          headers:
            X-RateLimit-Remaining:
              $ref: "#/components/headers/X-RateLimit-Remaining"
            X-RateLimit-Reset:
              $ref: "#/components/headers/X-RateLimit-Reset"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ChatCompletionResponse"
            text/event-stream:
              schema:
                $ref: "#/components/schemas/ChatCompletionChunk"
        "401":
          description: Unauthorized.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "429":
          description: Rate limit exceeded.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /v1/embeddings:
    post:
      operationId: openaiEmbeddings
      tags: [OpenAI]
      summary: Create embeddings
      description: Generate embedding vectors for the given input.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/EmbeddingRequest"
      responses:
        "200":
          description: Embedding result.
          headers:
            X-RateLimit-Remaining:
              $ref: "#/components/headers/X-RateLimit-Remaining"
            X-RateLimit-Reset:
              $ref: "#/components/headers/X-RateLimit-Reset"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EmbeddingResponse"
...